Make main executable compile shaders via the shader compiler
Rework Platform and introduce GraphicsApi, to disambiguate when passing parameters into the shader compiler
Make shader compiler work with UTF-8 on Windows by transforming the widechar command line to multibyte (Windows terminology). On other platforms it should just work
Make several enums have a smaller uint8_t type
Export a GlobalVariables.h (could be called GlobalPaths.h if need be)
Copy shader compiler to a known location and invoke the compiler to build the shaders

+++ Source/Rendering/CrRendering.h
@@ -326,7 +326,7 @@ namespace cr3d
 
 	namespace GraphicsApi
 	{
-		enum T : uint32_t
+		enum T : uint8_t
 		{
 			Vulkan,
 			D3D12,
@@ -337,7 +337,7 @@ namespace cr3d
 
 	namespace TextureContent
 	{
-		enum T : uint32_t
+		enum T : uint8_t
 		{
 			Diffuse,
 			Normals,
@@ -380,7 +380,7 @@ namespace cr3d
 	namespace ShaderStage
 	{
 		// Most shaders won't have Hull/Domain or Geometry, for performance we'll reorder them
-		enum T : uint32_t
+		enum T : uint8_t
 		{
 			Vertex,
 			Pixel,
@@ -400,7 +400,7 @@ namespace cr3d
 
 	namespace ResourceState
 	{
-		enum T : uint32_t
+		enum T : uint8_t
 		{
 			Undefined        = 0, // Never use this as the destination state in a resource transition operation
 			ShaderInput      = 1, // Use as input to a shader other than the pixel shader
